Escape table and column names and make them case-sensitive
Oracle names are case-insensitive in the way that Oracle always
translate them to upper case.
In order to preserve entity property names in the correct case without
employing some kind of translation we need to always quote them.
This presume that tables and columns were created with quoted names and
case in entity properties exactly match the column name.

See
http://docs.oracle.com/cd/B19306_01/server.102/b14200/sql_elements008.htm